corrections

+++ 0x1E-search_algorithms/1-binary.c
@@ -42,9 +42,9 @@ int binary_search(int *array, size_t size, int value)
  * @b: end of array
  */
 
-void print_array(int *array, int a, int b)
+void print_array(int *array, size_t a, size_t b)
 {
-	int i;
+	size_t i;
 
 	for (i = a; i <= b; i++)
 	{
