+++ 0x1E-search_algorithms/search_algos.h
@@ -3,6 +3,6 @@
 
 int linear_search(int *array, size_t size, int value);
 int binary_search(int *array, size_t size, int value);
-void print_array(int *array, int a, int b);
+void print_array(int *array, size_t a, size_t b);
 
 #endif
